finished divar crawler, bug fixes
