+++ crawler/utils.go
@@ -1,14 +1,33 @@
 package crawler
 
 import (
-	"fmt"
+	"encoding/json"
 	"log"
+	"os"
 	"regexp"
 	"strconv"
 	"strings"
 
 	"github.com/go-rod/rod"
 )
+
+type City struct {
+	Display string `json:"display"`
+	Slug    string `json:"slug"`
+}
+
+type District struct {
+	Display string `json:"display"`
+	Slug    string `json:"slug"`
+}
+
+type Cities struct {
+	Cities []City `json:"cities"`
+}
+
+type Districts struct {
+	Districts []District `json:"districts"`
+}
 
 func cleanPrices(a string) string {
 	return strings.ReplaceAll(strings.TrimSpace(strings.ReplaceAll(strings.ReplaceAll(a, `٬`, ``), `تومان`, ``)), `.`, ``)
@@ -47,8 +66,51 @@ func GetLatitudeAndLongitude(a string) (float64, float64) {
 func getNumbersFromSections(reg string, collector *rod.Page) string {
 	if ok, section, _ := collector.HasR(`div.kt-base-row.kt-base-row--large.kt-unexpandable-row`, reg); ok {
 		uncleaned := section.MustElement(`p.kt-unexpandable-row__value`).MustText()
-		fmt.Println(uncleaned, reg)
 		return uncleaned
 	}
 	return ""
+}
+
+func getDistricts() (Districts, error) {
+	file, err := os.Open("../database/districts.json")
+	if err != nil {
+		return Districts{}, err
+	}
+	defer file.Close()
+
+	data, err := os.ReadFile("../database/districts.json")
+	if err != nil {
+		return Districts{}, err
+	}
+
+	var districts Districts
+	err = json.Unmarshal(data, &districts)
+	if err != nil {
+		return Districts{}, err
+	}
+
+	return districts, nil
+
+}
+
+func getCities() (Cities, error) {
+	file, err := os.Open("./database/city.json")
+	if err != nil {
+		return Cities{}, err
+	}
+	defer file.Close()
+
+	data, err := os.ReadFile("./database/city.json")
+	if err != nil {
+		return Cities{}, err
+	}
+
+	var cities Cities
+	err = json.Unmarshal(data, &cities)
+	if err != nil {
+		return Cities{}, err
+	}
+
+	return cities, nil
+
 }
