--- models/models.go
@@ -1,9 +1,0 @@
-package models
-
-import "github.com/google/uuid"
-
-type Advertisements struct {
-	Id    uuid.UUID
-	title string
-	// TODO : add other fields
-}
